新規登録機能をAdminControllerへ移行
